Stage

+++ src/TD.Web/TD.Web.Common/TD.Web.Common.Contracts/Requests/Users/UserSaveLastTimeSeenRequest.cs
@@ -9,7 +9,7 @@ namespace TD.Web.Common.Contracts.Requests.Users
         public UserSaveLastTimeSeenRequest()
             : base() { }
 
-        public UserSaveLastTimeSeenRequest(int userId)
+        public UserSaveLastTimeSeenRequest(long userId)
             : base(userId) { }
     }
 }
